Rename my dream component
